==> YGG/3-wage-deductions.ipynb <==
import pandas as pd
import panel as pn
import numpy as np
pn.extension()

data = pd.read_csv('outputs/praise_data.csv')
total_hours = pd.read_csv('outputs/total_hours.csv')
receivers = pd.read_csv('outputs/receivers.csv')
quantifiers = pd.read_csv('outputs/quantifiers.csv')
received = receivers.groupby('To')[['IH per Praise']].sum().sort_values('IH per Praise', ascending=False).reset_index()
dished = receivers.groupby('From')[['IH per Praise']].sum().sort_values('IH per Praise', ascending=False).reset_index()
quantified = quantifiers.groupby('To')[['IH per Praise']].sum().sort_values('IH per Praise', ascending=True).reset_index()
net = data.groupby('To')[['IH per Praise']].sum().sort_values('IH per Praise', ascending=False).reset_index()

net = data.groupby('To')[['IH per Praise']].sum().sort_values('IH per Praise', ascending=False).reset_index()

dished = dished.rename({'IH per Praise': 'Total Hours Given'}, axis=1).set_index('From')

received = received.rename({'IH per Praise': 'Total Hours Received'}, axis=1).set_index('To')

quantified = quantified.rename({'IH per Praise': 'Total Hours Deducted'}, axis=1).set_index('To')

net = net.rename({'IH per Praise': 'Default Net Impact Hours'}, axis=1).set_index('To')

df = dished.merge(received, how='outer', left_index=True, right_index=True).merge(quantified, how='outer', left_index=True, right_index=True).merge(net, how='outer', left_index=True, right_index=True)

df['Initial Distribution'] = df['Default Net Impact Hours'] / df['Default Net Impact Hours'].sum()

df.sum()

df = df.sort_values('Default Net Impact Hours', ascending=False).replace(np.nan, 0)

df['40% Wage Restoration Augmentation'] = abs(-df['Total Hours Deducted'] * 0.4)

df.sum()

df.head(20)

df.to_csv('outputs/YGG-WAGE-AUGMENTATION.csv')
