==> YGG/4-proposal.ipynb <==
### Praise Catgories Augmentation
~~1. Category Distributions - Meetings and Tweet Augmentation~~

### Praise Distributions Augmentation

1. 40% Wage Reductions Returned to Workers

2. The default distribution is random. Not even normal random, but like pareto random, which is dangerous. UBI should be applied to smooth it. Power to the people. Wisdom of the crowd.

3. Gifting Economy - Rewarding Praise
Earn up to 25 bonus impact hours by giving out up to 100 impact hours of praise!

4. Hidden Work and Legacy Attribution - Decentralized Effort.  
This is the rockstar approach. Set a date and time. I'm going for Jeff Emmet, Angela, and Will R.

5. NFT Trophies
Top Hatcher. @zeptimus  
Highest Giver. @iviangita  
Most UBI.  

import pandas as pd

# Wage Augmentation
wage_augmentation = pd.read_csv('outputs/YGG-WAGE-AUGMENTATION.csv').set_index('Unnamed: 0')

wage_augmentation.head(40)

### UBI Augmentation

# UBI Augmentation
ubi_augmentation = pd.read_csv("outputs/YGG-UBI-Augmentation.csv").set_index('To')

# Merge
augmentation = wage_augmentation.merge(ubi_augmentation[['Augmentation']], left_index=True, right_index=True).rename({'Augmentation':'UBI Augmentation'}, axis=1)

### Gifting Economy Augmentation

augmentation['Gifting Economy Augmentation'] = (augmentation['Total Hours Given'] * 0.25).clip(upper=25)
augmentation = augmentation.reset_index().rename(columns={'index': 'Hatcher'})

### Composing the Final Distribution

augmentation['Composite Augmentations'] = augmentation[['40% Wage Restoration Augmentation', 'UBI Augmentation', 'Gifting Economy Augmentation']].sum(axis=1)
augmentation['Augmented Impact Hours'] = augmentation['Default Net Impact Hours'] + augmentation['Composite Augmentations']
augmentation['Final Distribution'] = augmentation['Augmented Impact Hours'] / augmentation['Augmented Impact Hours'].sum()
augmentation['Diff'] = augmentation['Final Distribution'] - augmentation['Initial Distribution']

augmentation = augmentation.sort_values("Final Distribution", ascending=False).reset_index()
augmentation.head(20)

import numpy as np

def resources_percentage(p, data, column):
    relevant_percentile = np.percentile(data[column],p)
    is_gt_relevant_percentile = data[column] > relevant_percentile
    filtered_data = data[is_gt_relevant_percentile]
    filtered_hours = filtered_data[column]
    pct_hours = filtered_hours.sum()/data[column].sum()
    return pct_hours

for p in [99, 95, 80, 50]:
    print("The top {} percent of the population".format(100-p))
    print("received {:0.2f} percent of the Impact Hours \n \n".format(resources_percentage(p, augmentation, 'Augmented Impact Hours')))

def gini_coefficient(x):
    n = len(x)
    x_bar = np.mean(x)
    abs_diffs = np.array([np.sum(np.abs(x[i] - x)) for i in range(n)])
    sum_abs_diffs = np.sum(abs_diffs)
    denominator = 2*n*n*x_bar
    return sum_abs_diffs/denominator

gini_coefficient(augmentation['Augmented Impact Hours'])

Max UBI Trophy

augmentation.reset_index().to_csv('outputs/YGG-AUGMENTATION.csv', index=False)
